Testing bug fix for IOSpeed.cc, changed nCellGlobal variable in workBound.c to long long unsigned to avoid potential overflow.
git-svn-id: svn+ssh://lc-svn.llnl.gov/usr/gapps/emhm/svn/Cardioid/EP/trunk@1151 baecf98e-1e81-4516-8920-d4c889e67ea4

+++ src/workBound.c
@@ -97,7 +97,7 @@ int  findChunks(BALANCER *balancer, unsigned short *seg, double diffCost, COLUMN
    }
    return offset; 
 }
-BALANCER buildBalancer(int nCellGlobal, int nx,int ny,int nz,int dx,int dy,int dz, 
+BALANCER buildBalancer(long long unsigned nCellGlobal, int nx,int ny,int nz,int dx,int dy,int dz, 
        int  nTasks, int nCores, int nRCoresBB, double alpha, double beta, int printStats, MPI_Comm comm) 
 {
    int rank,nRank;
